onconova-server/pyproject.toml
[tool.poetry]
name = "onconova-server"
version = "0.4.0"
description = "The Precision Oncology Program (Onconova) is a joint at the University Hospital Zurich (USZ) to capture clinical data from oncology patients in a structured manner. The final aim of the project is to enable aggregated data analysis interactively."
readme = "../README.md"
license = "MIT"
authors = ["Luis Fabregas-Ibanez <mail@luisfabregas.com>"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]
packages = [{include = "onconova"}]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
# Core dependencies
python = "~=3.11"
django = "5.1"
psycopg = "3.2.3"
psycopg2 = "2.9.10"
# Django extensions/apps
django_extensions = "3.2.1"
django-allauth = "65.8.0"
django-ninja-extra = "0.22.3"
django-cors-headers = "4.6.0"
django-queryable-properties = "1.10.0"
django-pghistory = "3.5.5"
django-stubs = "5.2.2"
django-stubs-ext = "5.2.2"
# Tools
measurement = "3.2.2"
requests = "2.32.0"
requests-oauthlib = "2.0.0"
cachetools = "5.5.0"
tqdm = "4.67.0"
jwt = "1.3.1"
xlsx2csv = "0.8.4"
pyyaml = "6.0.2"
# Testing
pytest-django = "4.5.2"
factory_boy = "3.3.1"
parameterized = "0.9.0"
coverage = "7.2.3"
# WSGI server for production
gunicorn = "21.2.0"
# Serve dev-instance via HTTPS
Werkzeug = "3.0.0"
pyOpenSSL = "23.2.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "onconova.tests.settings"
python_files = [
    "onconova/tests/**/test_*.py",
]
testpaths = ['onconova/tests']
addopts = "--no-migrations --reuse-db"
django_find_project = "True"
django_debug_mode = "True"

[tool.coverage.run]
source = ["."]
omit = [
  "./.venv/*",
  "./venv/*",
  "*test*",
  "test_*",
  "*apps.py",
  "*manage.py",
  "*conftest.py",
  "*__init__.py",
  "*migrations*",
  "*asgi*",
  "*wsgi*",
  "*admin.py",
  "*urls.py",
  "**/migrations/*"
]

[tool.coverage.report]
omit = [
  "./.venv/*",
  "./venv/*",
  "*test*",
  "*apps.py",
  "*manage.py",
  "*conftest.py",
  "*__init__.py",
  "*migrations*",
  "*asgi*",
  "*wsgi*",
  "*admin.py",
  "*urls.py",
  "**/migrations/*"
]

[tool.isort]
skip = ["__init__.py"]
profile = "black"